--- src/assets/scss/iconfont.scss
@@ -1,2 +1,0 @@
-// TODO 开发环境和测试分开
-@import '//at.alicdn.com/t/c/font_4039139_gyojkkwjz0l.css';
