+++ src/assets/scss/index.scss
@@ -1,4 +1,2 @@
 // 全局样式，被 nuxt.config.ts 全局引入
-@use './reset';
-@use './iconfont';
 @use './transitions';
