--- src/assets/scss/reset.scss
@@ -1,12 +1,0 @@
-// 全局的重置样式
-html,
-body,
-#app {
-  padding: 0;
-  margin: 0;
-}
-
-a {
-  font-weight: normal;
-  color: var(--el-color-primary);
-}
