+++ src/layouts/default.vue
@@ -1,6 +1,7 @@
 <script lang="ts" setup>
 import zh from 'element-plus/dist/locale/zh-cn.mjs'
 import en from 'element-plus/dist/locale/en.mjs'
+import AppHeader from './components/AppHeader.vue'
 
 const { locale } = useI18n()
 
@@ -11,7 +12,7 @@ const localeData = computed(() => (locale.value === 'en' ? en : zh))
   <el-config-provider :locale="localeData">
     <el-container class="min-h-screen">
       <el-header>
-        <LayoutHeader />
+        <AppHeader />
       </el-header>
       <el-main>
         <slot />
